README.md
This repository is about the game called Hangman. Hangman is a multiplayer game where one player selects a word, and other players have a certain number of guesses to guess the characters in the word. The objective is to guess the entire word within a specified number of attempts to win the game. If the players fail to guess the word within the given attempts, they lose the game.